requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
psycopg2-binary==2.9.9
requests==2.32.3
urllib3==2.2.2
# apache-airflow==2.9.3
# apache-airflow-providers-docker==3.12.2